feat(types): add support for NpcType and NpcSubType
Add support for NpcType and NpcSubType in the ITypeService interface and its implementation. Also, add support for adding NpcType and NpcSubType in the TypesScriptModule. Finally, add support for parsing NpcType and NpcSubType from a CSV file in the SeedService.

feat(TypeService.cs): add support for NpcType and NpcSubType
This commit adds support for NpcType and NpcSubType to the TypeService class. It includes the addition of two new lists, _npcTypes and _npcSubTypes, and the creation of methods to add and retrieve NpcTypes and NpcSubTypes. The methods include AddNpcType, AddNpcSubType, GetNpcType, and AddNpcSubType.