feat(network): add SignalR network client implementation
This commit adds a new SignalR network client implementation to the project. The client is added to the DI container in the `Program.cs` file. The client is configured with a new `DarkStarNetworkClientConfig` object that specifies the address and port of the SignalR server. The `SignalrNetworkClient` class is added to the `DarkStar.Network.Client` namespace. The client is used to send and receive messages to and from the server. The `SignalrNetworkServer` class is updated to use the `sessionId` parameter as a string instead of an object. The `SendMessageAsync` method is updated to use the `sessionId` parameter as a string instead of an object. The `OnMessageReceived` event is updated to check if it is null before invoking it.

refactor(JsonMessageBuilder.cs): replace System.Text.Json with Newtonsoft.Json for deserialization
refactor(AccountCreateRequestMessage.cs): remove unnecessary null assignment in Email and Password properties